+++ addons/main/CfgMagazineWells.hpp
@@ -1,6 +1,6 @@
 class CfgMagazineWells {
     class AK_762x39 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_ak47mag_NT",
             "uns_ak47mag",
             "uns_ak47mag_T"
@@ -8,35 +8,35 @@ class CfgMagazineWells {
     };
 
     class CBA_12g_2rnds {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_12gaugemag_2",
             "uns_12gaugemag_2f"
         };
     };
 
     class CBA_12g_4rnds {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_12gaugemag_4",
             "uns_12gaugemag_4f"
         };
     };
 
     class CBA_12g_6rnds {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_12gaugemag_6",
             "uns_12gaugemag_6f"
         };
     };
 
     class CBA_12g_8rnds {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m870mag",
             "uns_m870mag_8f"
         };
     };
 
     class CBA_22mm_GL_M1 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_1Rnd_22mm_lume",
             "Uns_1Rnd_22mm_FRAG",
             "Uns_1Rnd_22mm_WP",
@@ -48,13 +48,13 @@ class CfgMagazineWells {
     };
 
     class CBA_30mm_GL_Type99 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "Uns_1Rnd_30mm_FRAG"
         };
     };
 
     class CBA_3006_BAR {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_barmag_NT",
             "uns_barmag",
             "uns_barmag_T"
@@ -62,41 +62,41 @@ class CfgMagazineWells {
     };
 
     class CBA_3006_Belt {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_100Rnd_m1919"
         };
     };
 
     class CBA_3006_Garand {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m1garandmag",
             "uns_m1garandmag_T"
         };
     };
 
     class CBA_3006_Model70 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_model70mag",
             "uns_model70mag_T"
         };
     };
 
     class CBA_3006_Spring {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_springfieldmag",
             "uns_springfieldmag_T"
         };
     };
 
     class CBA_303B_LeeEn {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_smlemag",
             "uns_smlemag_T"
         };
     };
 
     class CBA_30Carbine_M1Carbine {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m1carbinemag_NT",
             "uns_m1carbinemag",
             "uns_m1carbinemag_T",
@@ -107,14 +107,14 @@ class CfgMagazineWells {
     };
 
     class CBA_32ACP_Vz61 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_20Rnd_sa61",
             "uns_20Rnd_sa61_pa"
         };
     };
 
     class CBA_40mm_M203 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_1Rnd_AB_M397",
             "uns_1Rnd_BS_M576",
             "uns_40mm_green",
@@ -140,14 +140,14 @@ class CfgMagazineWells {
     };
 
     class CBA_45ACP_Grease {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m3a1mag",
             "uns_m3a1mag_T"
         };
     };
 
     class CBA_45ACP_Thompson_Stick {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_thompsonmag_20",
             "uns_thompsonmag_20_NT",
             "uns_thompsonmag_20_T",
@@ -158,28 +158,28 @@ class CfgMagazineWells {
     };
 
     class CBA_556x45_HK33 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_40Rnd_556x45_t223_T",
             "uns_40Rnd_556x45_t223"
         };
     };
 
     class CBA_75x54mmFrench_MAS36 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_mas36mag",
             "uns_mas36mag_T"
         };
     };
 
     class CBA_75x54mmFrench_MAS49 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_mas4956mag",
             "uns_mas4956mag_T"
         };
     };
 
     class CBA_762x25_MAT49 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_mat49_20_mag",
             "uns_mat49mag_NT",
             "uns_mat49mag",
@@ -188,7 +188,7 @@ class CfgMagazineWells {
     };
 
     class CBA_762x25_PPSh_Drum {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_ppshmag_NT",
             "uns_ppshmag",
             "uns_ppshmag_T"
@@ -196,7 +196,7 @@ class CfgMagazineWells {
     };
 
     class CBA_762x25_PPSh_Stick {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_k50mag_NT",
             "uns_k50mag",
             "uns_k50mag_T"
@@ -204,7 +204,7 @@ class CfgMagazineWells {
     };
 
     class CBA_762x39_AK {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_ak47mag_NT",
             "uns_ak47mag",
             "uns_ak47mag_T"
@@ -212,21 +212,21 @@ class CfgMagazineWells {
     };
 
     class CBA_762x39_RPD {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_rpdmag",
             "uns_rpdsogmag"
         };
     };
 
     class CBA_762x39_RPK {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_rpkmag",
             "uns_75Rnd_akdr"
         };
     };
 
     class CBA_762x39_SKS {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_sksmag_NT",
             "uns_sksmag",
             "uns_sksmag_T"
@@ -234,7 +234,7 @@ class CfgMagazineWells {
     };
 
     class CBA_762x39_SKS_Clip {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_sksmag_NT",
             "uns_sksmag",
             "uns_sksmag_T"
@@ -242,7 +242,7 @@ class CfgMagazineWells {
     };
 
     class CBA_762x39_VZ58 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_sa58mag_NT",
             "uns_sa58mag",
             "uns_sa58mag_T"
@@ -250,14 +250,14 @@ class CfgMagazineWells {
     };
 
     class CBA_762x51_5rnds {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m40mag",
             "uns_m40mag_T"
         };
     };
 
     class CBA_762x51_FAL {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_l1a1mag_NT",
             "uns_l1a1mag",
             "uns_l1a1mag_T"
@@ -265,14 +265,14 @@ class CfgMagazineWells {
     };
 
     class CBA_762x51_LINKS {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m60mag",
             "uns_m60mag_200"
         };
     };
 
     class CBA_762x51_M14 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m14mag_NT",
             "uns_m14mag",
             "uns_m14mag_T"
@@ -280,73 +280,73 @@ class CfgMagazineWells {
     };
 
     class CBA_762x54R_DPM {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_47Rnd_DP28"
         };
     };
 
     class CBA_762x54R_LINKS {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_100Rnd_762x54_PK"
         };
     };
 
     class CBA_762x54R_Mosin {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_mosinmag",
             "uns_mosinmag_T"
         };
     };
 
     class CBA_762x54R_SVD {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_svdmag",
             "uns_svdmag_T"
         };
     };
 
     class CBA_762x54R_SVT {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_svtmag",
             "uns_svtmag_T"
         };
     };
 
     class CBA_762x54R_Vz59_LINKS {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_100Rnd_762x54_ukvz59"
         };
     };
 
     class CBA_77x58_Arisaka {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_type99mag",
             "uns_type99mag_T"
         };
     };
 
     class CBA_792x33_StG {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_30Rnd_kurtz_stg"
         };
     };
 
     class CBA_792x57_K98 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_kar98kmag",
             "uns_kar98kmag_T"
         };
     };
 
     class CBA_792x57_LINKS {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_50Rnd_792x57_Mg42",
             "uns_250Rnd_792x57_Mg42"
         };
     };
 
     class CBA_8x22_Type100 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_type100_NT",
             "uns_type100mag",
             "uns_type100_T"
@@ -354,21 +354,21 @@ class CfgMagazineWells {
     };
 
     class CBA_9x18_PM63 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_25Rnd_pm",
             "uns_25Rnd_pm_pa"
         };
     };
 
     class CBA_9x19_Owen {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_owenmag",
             "uns_owenmag_T"
         };
     };
 
     class CBA_9x19_M45CarlG {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m45mag_NT",
             "uns_m45mag",
             "uns_m45mag_T"
@@ -376,25 +376,25 @@ class CfgMagazineWells {
     };
 
     class CBA_9x19_M50Madsen {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m50mag"
         };
     };
 
     class CBA_9x19_MAC10 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_mac10mag"
         };
     };
 
     class CBA_9x19_MP40 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_mp40mag"
         };
     };
 
     class CBA_9x19_STEN {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_stenmag",
             "uns_stenmag_T",
             "uns_mk4mag",
@@ -404,53 +404,53 @@ class CfgMagazineWells {
     };
 
     class CBA_9x19_UZI {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_32Rnd_uzi",
             "uns_32Rnd_uzi_pa"
         };
     };
 
     class CBA_Bazooka {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_M28A2_mag",
             "uns_M30_smoke_mag"
         };
     };
 
     class CBA_RPG7 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_rpg7grenade"
         };
     };
 
     class M240_762x51 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_m60mag",
             "uns_m60mag_200"
         };
     };
 
     class PK_762x54R {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_100Rnd_762x54_PK"
         };
     };
 
     class RPG7 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_rpg7grenade"
         };
     };
 
     class SVD_762x54R {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_svdmag",
             "uns_svdmag_T"
         };
     };
 
     class UGL_40x36 {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_1Rnd_AB_M397",
             "uns_1Rnd_BS_M576",
             "uns_40mm_green",
@@ -476,7 +476,7 @@ class CfgMagazineWells {
     };
 
     class UNS_556x45_STANAG {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_20Rnd_556x45_Stanag_NT",
             "uns_20Rnd_556x45_Stanag_T",
             "uns_20Rnd_556x45_Stanag",
@@ -493,7 +493,7 @@ class CfgMagazineWells {
     };
 
     class UNS_762x25_PPSh {
-        WB_Magazines[] += {
+        MAGAZINE_WELL[] += {
             "uns_ppshmag_NT",
             "uns_ppshmag",
             "uns_ppshmag_T",