+++ addons/main/script_mod.hpp
@@ -3,11 +3,18 @@
 
 #include "script_version.hpp"
 
-#define VERSION     MAJOR.MINOR.PATCHLVL.BUILD
+#define VERSION     MAJOR.MINOR
 #define VERSION_STR MAJOR.MINOR.PATCHLVL.BUILD
 #define VERSION_AR  MAJOR,MINOR,PATCHLVL,BUILD
 
 // MINIMAL required version for the Mod. Components can specify others..
-#define REQUIRED_VERSION 2.06
+#define REQUIRED_VERSION 2.14
 
-#define COMPONENT_NAME QUOTE(Weapons Balance - Unsung)
+#ifndef COMPONENT_BEAUTIFIED
+    #define COMPONENT_BEAUTIFIED COMPONENT
+#endif
+#ifdef SUBCOMPONENT_BEAUTIFIED
+    #define COMPONENT_NAME QUOTE(Weapons Balance - COMPONENT_BEAUTIFIED - SUBCOMPONENT_BEAUTIFIED)
+#else
+    #define COMPONENT_NAME QUOTE(Weapons Balance - COMPONENT_BEAUTIFIED)
+#endif