+++ addons/main/weapons/CfgWeapAR15.hpp
@@ -9,29 +9,21 @@ class uns_m16: Uns_Rifle {
 };
 class uns_m16_camo: uns_m16 {
     displayName = "M16 (Camo)";
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
+    class WeaponSlotsInfo;
 };
 class uns_M16_dm: uns_m16 {
     magazineReloadTime = 0;
     scope = 1;
     scopeArsenal = 1;
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 class uns_m16_m203: uns_m16 {
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 101.5;
     };
 };
-class uns_m16_xm148: uns_m16_m203 {
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
-};
+class uns_m16_xm148;
 class uns_m16_xm148_camo: uns_m16_xm148 {
     displayName = "M16 (Camo/XM148)";
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
-};
-
-class uns_M16_NV: uns_m16 {
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 
 class uns_M16A1: uns_m16 {
@@ -41,10 +33,9 @@ class uns_M16A1: uns_m16 {
 };
 class uns_m16A1_camo: uns_M16A1 {
     displayName = "M16A1 (Camo)";
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 class uns_M16A1_HBAR_base: uns_M16A1 {
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
+    class WeaponSlotsInfo;
 };
 class uns_M16A1_HBAR: uns_M16A1_HBAR_base {
     class WeaponSlotsInfo: WeaponSlotsInfo {
@@ -56,13 +47,6 @@ class uns_m16a1_m203: uns_M16A1 {
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 101.5;
     };
-};
-class uns_m16a1_xm148: uns_m16a1_m203 {
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
-};
-
-class uns_M16A1_NV: uns_M16A1 {
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 
 class uns_xm177e1: uns_m16 {
@@ -76,7 +60,6 @@ class uns_xm177e1_dm: uns_xm177e1 {
     displayName = "XM177E1 (Dual Mag)";
     scope = 1;
     scopeArsenal = 0;
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 
 class uns_xm177e1_m203: uns_m16_m203 {
@@ -108,7 +91,7 @@ class uns_xm177e2_xm148: uns_xm177e2_m203 {
 class uns_xm177e2: uns_xm177e1 {
     ACE_barrelLength = 292.1;
     displayName = "XM177E2";
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
+    class WeaponSlotsInfo;
 };
 class uns_xm177e2_grip: uns_xm177e2 {
     displayName = "XM177E2 (Grip)";
@@ -125,7 +108,6 @@ class uns_xm177e2_stock: uns_xm177e2 {
 class uns_xm177e2_short: uns_xm177e2 {
     ACE_barrelLength = 254;
     displayName = "XM177E2 (Short)";
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 
 //https://en.wikipedia.org/wiki/Stoner_63
@@ -133,7 +115,7 @@ class uns_m63a_base: Uns_Rifle {
     ACE_barrelLength = 508;
     ACE_barrelTwist = 230;
     magazineReloadTime = 0;
-    class WeaponSlotsInfo: WeaponSlotsInfo {};
+    class WeaponSlotsInfo;
 };
 class uns_M63a_AR_base: uns_m63a_base {
     displayName = "Stoner 63A Rifle";